GuidedTours: Introduce continueIf condition for steps
Sometimes we want an ActionStep to be completed when a certain condition
has been reached, e.g. there are theme results found from a search.

We can specify this in a step config like:
```continueIf: ( state ) => themeResultsShown( state )```

The way is this implemented at the moment is suboptimal, as we are
passing the entire Redux state to the GuidedTours component, since we
want to call the `proceedToNextStep()` action, to make sure step
progression is correct in the state.

Middleware seems like a more reasonable place to do this, but should be
part of a later refactoring, when we're satifisied that the whole
GuidedTour framework… works.